out/pyproject.toml
[project]
name = "vequinox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.75.0",
    "cairosvg>=2.8.2",
    "datasets>=4.4.1",
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "llm-blender>=0.0.2",
    "matplotlib>=3.10.7",
    "peft>=0.15.0",
    "pillow>=12.0.0",
    "reportlab>=4.4.5",
    "streamlit>=1.51.0",
    "svglib>=1.6.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "trl[judges]>=0.25.1",
    "wandb>=0.23.0",
    "streamlit-drawable-canvas>=0.9.3",
]
